backend/requirements.txt
anyio==4.2.0 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
asgiref==3.7.2 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
asttokens==2.4.1 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
atomicwrites==1.4.1 ; python_full_version >= "3.11.5" and python_version < "4" and sys_platform == "win32"
attrs==23.2.0 ; python_full_version >= "3.11.5" and python_version < "4"
black==23.12.1 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
certifi==2024.2.2 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
cfgv==3.4.0 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
click==8.1.7 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
colorama==0.4.6 ; python_full_version >= "3.11.5" and python_version < "4" and (sys_platform == "win32" or platform_system == "Windows")
decorator==5.1.1 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
distlib==0.3.8 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
django-cors-headers==4.3.1 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
django-debug-toolbar==4.1.0 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
django-extensions==3.2.3 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
django==4.2.10 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
djangorestframework==3.14.0 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
drf-spectacular==0.26.5 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
executing==2.0.1 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
factory-boy==3.3.0 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
faker==23.1.0 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
filelock==3.13.1 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
flake8-docstrings==1.7.0 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
flake8==6.1.0 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
gunicorn==20.1.0 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
h11==0.14.0 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
httpcore==1.0.2 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
httpx==0.26.0 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
identify==2.5.34 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
idna==3.6 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
inflection==0.5.1 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
iniconfig==2.0.0 ; python_full_version >= "3.11.5" and python_version < "4"
ipython==8.21.0 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
isort==5.13.2 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
jedi==0.19.1 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
jsonschema-specifications==2023.12.1 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
jsonschema==4.21.1 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
matplotlib-inline==0.1.6 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
mccabe==0.7.0 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
mypy-extensions==1.0.0 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
nodeenv==1.8.0 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
packaging==23.2 ; python_full_version >= "3.11.5" and python_version < "4"
parso==0.8.3 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
pathspec==0.12.1 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
pexpect==4.9.0 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0" and sys_platform != "win32"
platformdirs==4.2.0 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
pluggy==1.4.0 ; python_full_version >= "3.11.5" and python_version < "4"
pre-commit==2.21.0 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
prompt-toolkit==3.0.43 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
psycopg2-binary==2.9.9 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
ptyprocess==0.7.0 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0" and sys_platform != "win32"
pure-eval==0.2.2 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
py==1.11.0 ; python_full_version >= "3.11.5" and python_version < "4"
pycodestyle==2.11.1 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
pydocstyle==6.3.0 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
pyflakes==3.1.0 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
pygments==2.17.2 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
pytest-deadfixtures==2.2.1 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
pytest-django==4.5.2 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
pytest-lazy-fixture==0.6.3 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
pytest-pythonpath==0.7.4 ; python_full_version >= "3.11.5" and python_version < "4"
pytest-sugar==0.9.7 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
pytest==6.2.5 ; python_full_version >= "3.11.5" and python_version < "4"
python-dateutil==2.8.2 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
python-dotenv==1.0.1 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
python-slugify==8.0.4 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
pytz==2024.1 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
pyyaml==6.0.1 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
referencing==0.33.0 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
rpds-py==0.17.1 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
setuptools==69.1.0 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
six==1.16.0 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
sniffio==1.3.0 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
snowballstemmer==2.2.0 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
sqlparse==0.4.4 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
stack-data==0.6.3 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
termcolor==2.4.0 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
text-unidecode==1.3 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
toml==0.10.2 ; python_full_version >= "3.11.5" and python_version < "4"
traitlets==5.14.1 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
tzdata==2024.1 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0" and sys_platform == "win32"
uritemplate==4.1.1 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
virtualenv==20.25.0 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
wcwidth==0.2.13 ; python_full_version >= "3.11.5" and python_full_version < "4.0.0"
